split tests into three test files and a utils file
